feat: Added handleCapyUrlCompany.ts
