+++ site/src/screens/Departments.tsx
@@ -1,7 +1,7 @@
 import useSWR from "swr";
 import ReactJson from "react-json-view";
-import toast from "react-hot-toast";
 
+import {handleCopyUrlCompany} from "../utils/handleCapyUrlCompany.ts";
 import Layout from "../components/Layout";
 import axios from "../config/axios";
 import Search from "../components/Search";
@@ -13,26 +13,6 @@ export default function Departments() {
     "/api/departamentos",
     fetchData
   );
-
-  const handleCopyUrlCompany = async () => {
-    const textElement = document.getElementById("urlEndpoint");
-
-    if (textElement) {
-      const text = textElement.textContent;
-      if (text) {
-        await navigator.clipboard.writeText(text);
-        toast("Enlace copiado", {
-          icon: "📋",
-        });
-      } else {
-        console.error("El contenido de texto está vacío.");
-      }
-    } else {
-      console.error(
-        "No se encontró el elemento con el ID 'urlCompany'."
-      );
-    }
-  };
 
   const handleSearch = () => {};
